AS-3092: Implement ImportCRAService.

+++ farms-api/farms-persistence/src/main/java/ca/bc/gov/farms/persistence/v1/dao/StagingDao.java
@@ -53,7 +53,7 @@ public interface StagingDao extends Serializable {
 
     void insert(final Z99ExtractFileCtl obj, final String userId) throws SQLException;
 
-    void status(Integer pImportVersionId, String pString) throws SQLException;
+    void status(Long pImportVersionId, String pString) throws SQLException;
 
-    void statusNonAutonomous(Integer pImportVersionId, String pString) throws SQLException;
+    void statusNonAutonomous(Long pImportVersionId, String pString) throws SQLException;
 }
